Add combatant script to act as a central manager of combat events

+++ Assets/GameplayAbilities/Runtime/ModificationRules/IAttributeClampRule.cs
@@ -2,7 +2,7 @@ using GameplayAbilities.Runtime.Attributes;
 
 namespace GameplayAbilities.Runtime.ModificationRules {
     public interface IAttributeClampRule {
-        public float MaxValueIn(AttributeSet root);
-        public float MinValueIn(AttributeSet root);
+        public int MaxValueIn(AttributeSet root);
+        public int MinValueIn(AttributeSet root);
     }
 }
